Support Execute Procedure Expect Exception in StandAlone mode
